backend/requirements.txt
alembic==1.14.1
Authlib==1.3.2
backports-datetime-fromisoformat==2.0.3
bcrypt==4.3.0
blinker==1.8.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
Flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0  # Changed from 4.7.1
flask-marshmallow==1.2.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==2.1.5
marshmallow==3.20.1
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.6.0
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.40
typing_extensions==4.13.2
urllib3==2.2.3
Werkzeug==3.0.6
